my_planner_bot/requirements.txt
aiogram==3.4.1
openai>=1.50.0
aiosqlite==0.20.0
apscheduler==3.10.4
python-dotenv==1.0.1